# Fix slow performance on forms#index